# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Pénitentiaire API",
        "description": "API REST pour le recensement des détenus, le suivi des condamnations et les statistiques pénitentiaires.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/detenus/{detenu}/condamnations": {
            "get": {
                "tags": [
                    "Condamnations"
                ],
                "summary": "Lister les condamnations d'un détenu",
                "operationId": "5c03d4842f7baba3c9e8882cd3df9d7e",
                "parameters": [
                    {
                        "name": "detenu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NINA-2026001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des condamnations"
                    },
                    "404": {
                        "description": "Détenu introuvable"
                    }
                }
            },
            "post": {
                "tags": [
                    "Condamnations"
                ],
                "summary": "Enregistrer une condamnation",
                "description": "La date de fin de peine est calculée automatiquement (date_debut + duree_peine_mois).",
                "operationId": "f348fbb6b48e960b2a2be56ef5a252d6",
                "parameters": [
                    {
                        "name": "detenu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NINA-2026001"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CondamnationInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Condamnation créée"
                    },
                    "404": {
                        "description": "Détenu introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/condamnations/{condamnation}": {
            "get": {
                "tags": [
                    "Condamnations"
                ],
                "summary": "Consulter une condamnation",
                "operationId": "5471949b562c0ff8801378f78d6c3e08",
                "parameters": [
                    {
                        "name": "condamnation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la condamnation"
                    },
                    "404": {
                        "description": "Condamnation introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Condamnations"
                ],
                "summary": "Modifier une condamnation",
                "operationId": "5291847d8b2d310f509f9b0057a05ffd",
                "parameters": [
                    {
                        "name": "condamnation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CondamnationInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Condamnation mise à jour"
                    },
                    "404": {
                        "description": "Condamnation introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Condamnations"
                ],
                "summary": "Supprimer une condamnation",
                "operationId": "a55b731c658c295366ca59034976af86",
                "parameters": [
                    {
                        "name": "condamnation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Condamnation supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Condamnation introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CondamnationInput": {
                "required": [
                    "id_infraction",
                    "id_juridiction",
                    "date_debut_peine",
                    "duree_peine_mois"
                ],
                "properties": {
                    "id_infraction": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_juridiction": {
                        "type": "integer",
                        "example": 1
                    },
                    "date_debut_peine": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-10"
                    },
                    "duree_peine_mois": {
                        "type": "integer",
                        "example": 36
                    }
                },
                "type": "object"
            },
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Opération réussie."
                    }
                },
                "type": "object"
            }
        }
    },
...
}